Desktop/energy-reader/backend/requirements.txt
# Django
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
qrcode==7.4.2

# API Documentation
drf-spectacular==0.27.0

# Celery
celery==5.3.4
redis==5.0.1

# OCR/Image Processing
pytesseract==0.3.10
paddlepaddle==2.5.2
paddleocr==2.7.3
opencv-python==4.8.1.78
Pillow==10.1.0
pdf2image==1.16.3
pyzbar==0.1.9

# Utilities
python-decouple==3.8
cryptography==41.0.8
python-magic==0.4.27
requests==2.31.0

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0